# apps/cas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CAS",
    "description": "The passkey identity provider of MemeBattle: the JSON API under `/api` for its own frontend, and the OpenID Connect endpoints at the root. Every error response lists the stable codes it can carry in `x-error-codes`; clients branch on those codes, never on messages.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/passkeys": {
      "get": {
        "summary": "The account's passkeys.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyListResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "unauthenticated"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "description": "For the developer; clients branch on `code`."
                        }
                      }
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "unauthenticated"
            ]
          },
          "500": {
            "description": "internal_error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "internal_error"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "description": "For the developer; clients branch on `code`."
                        }
                      }
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "internal_error"
            ]
          },
          "503": {
            "description": "database_busy, database_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "database_busy",
                            "database_unavailable"
                          ]
                        },
                        "message": {
                          "type": "string",
                          "description": "For the developer; clients branch on `code`."
                        }
                      }
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "database_busy",
              "database_unavailable"
            ]
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PasskeyListResponse": {
        "type": "object",
        "required": [
          "passkeys"
        ],
        "properties": {
          "passkeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PasskeyResponse"
            }
          }
        }
      },
      "PasskeyResponse": {
        "type": "object",
        "description": "A passkey as the dashboard shows it. The credential itself (public key,\ncounter, flags) is server-side state and never leaves.",
        "required": [
          "id",
          "name",
          "createdAt",
          "lastUsedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "`null` until the passkey is first used to sign in."
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic",
        "description": "`client_secret_basic`: a confidential client's id and secret."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "An access token `/token` issued (ADR 0011)."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cas_session",
        "description": "The session cookie set by registration and login (ADR 0004). On an https origin its name is `__Host-cas_session`."
      }
    }
  }
}